requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
openai>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
websockets>=11.0.0
flask-socketio>=5.3.0
python-socketio>=5.9.0
pydub>=0.25.1
google-cloud-speech>=2.21.0
requests-oauthlib>=1.3.1
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dateutil>=2.8.2 